Move SpotsDelegate to Shared
